#167 Moved YamlReader to Core module, dependencies cleanup, Context removed
