+++ bb-core/pom.xml
@@ -80,6 +80,11 @@ limitations under the License.
             <groupId>org.assertj</groupId>
             <artifactId>assertj-core</artifactId>
         </dependency>
+        <dependency>
+            <groupId>com.fasterxml.jackson.dataformat</groupId>
+            <artifactId>jackson-dataformat-yaml</artifactId>
+            <version>2.5.4</version>
+        </dependency>
 
         <!-- logging -->
         <dependency>
