+++ bb-core/src/main/java/com/cognifide/qa/bb/utils/config/YamlReader.java
@@ -17,7 +17,7 @@
  * limitations under the License.
  * #L%
  */
-package com.cognifide.qa.bb.aem.touch.util;
+package com.cognifide.qa.bb.utils.config;
 
 import java.io.File;
 import java.io.IOException;
@@ -60,7 +60,7 @@ public final class YamlReader {
       return mapper.readValue(file, typeReference);
     } catch (IOException | URISyntaxException e) {
       LOG.error("Could not read YAML file: {}", path);
-      throw new IllegalStateException("YAML file could not be read");
+      throw new IllegalStateException("YAML file could not be read", e);
     }
   }
 }
